Rename faultTypes to types
