--- src/main/java/cloud/filibuster/junit/configuration/examples/redis/byzantine/faultTypes/ByzantineFault.java
@@ -1,5 +1,0 @@
-package cloud.filibuster.junit.configuration.examples.redis.byzantine.faultTypes;
-
-enum ByzantineFault {
-    STRING, BYTE_ARRAY
-}
