+++ src/main/java/cloud/filibuster/junit/configuration/examples/redis/byzantine/types/ByzantineStringFaultType.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.junit.configuration.examples.redis.byzantine.faultTypes;
+package cloud.filibuster.junit.configuration.examples.redis.byzantine.types;
 
 import javax.annotation.Nullable;
